add startup flag args ," -f " that set config path for hosting.
